+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Changed
 
-- Code folding is now conscious of spec login sessions (`welcome to "spec" Release x.x.x` line)
+- Code folding and navigation features are now conscious of spec sessions (`welcome to "spec" Release x.x.x` line)
   - prompt lines such as `1.SPEC> ...` are grouped in the respective sessions.
   - Previously it was a flat sctucture consisting of the prompt lines.